refactor(ui): rewrite cursor page to design system (Phase 4) (#1101)
* refactor(ui): rewrite cursor page to design system

Phase 4 outlier rewrite — 1413 LOC monolith split into:
- ui/src/pages/cursor/index.tsx (root, 40 LOC)
- ui/src/pages/cursor/cursor-form.tsx (state + handlers)
- ui/src/pages/cursor/sections/status-sidebar.tsx
- ui/src/pages/cursor/sections/presets-section.tsx
- ui/src/pages/cursor/sections/model-mapping-section.tsx
- ui/src/pages/cursor/sections/runtime-settings-section.tsx
- ui/src/pages/cursor/sections/info-section.tsx
- ui/src/pages/cursor/sections/manual-auth-dialog.tsx
- ui/src/pages/cursor/lib/cursor-config-helpers.ts

Replaces Tabs (Model Config / Settings / Info) with FormSections
in FormPane + SectionRail click-to-scroll nav. All flows preserved:
probe, auto-detect auth, manual auth dialog, daemon start/stop,
model presets, raw JSON editor, enable/disable toggle.

Updates height-contract test path, vitest-setup IntersectionObserver
stub, and pre-existing cursor-page integration test.

* docs(ui): add force-UI PR materials for cursor rewrite

- Before/after screenshots at 1440x900 with privacy mode
- Self-contained HTML report with base64 screenshots, LOC delta,
  catalogue checklist, validation badges
- 43-test cursor-page.test.tsx covering all sections, dialogs, flows
