+++ ui/tests/setup/vitest-setup.ts
@@ -36,6 +36,18 @@ class ResizeObserverMock {
 }
 global.ResizeObserver = ResizeObserverMock as unknown as typeof ResizeObserver;
 
+// Mock IntersectionObserver — required by SectionRail (config-layout) scroll-spy
+class IntersectionObserverMock {
+  root: Element | Document | null = null;
+  rootMargin = '';
+  thresholds: ReadonlyArray<number> = [];
+  observe = vi.fn();
+  unobserve = vi.fn();
+  disconnect = vi.fn();
+  takeRecords = vi.fn(() => []);
+}
+global.IntersectionObserver = IntersectionObserverMock as unknown as typeof IntersectionObserver;
+
 // Mock localStorage
 const localStorageMock = {
   getItem: vi.fn(),
